feat(storages): support large keys and values

+++ src/const.h
@@ -726,8 +726,8 @@ static constexpr int32_t CHANNEL_PRIVATE = 0xFFFF;
 
 // Reserved player storage key ranges;
 //[10000000 - 20000000];
-static constexpr int32_t PSTRG_RESERVED_RANGE_START = 10000000;
-static constexpr int32_t PSTRG_RESERVED_RANGE_SIZE = 10000000;
+static constexpr int64_t PSTRG_RESERVED_RANGE_START = 10000000;
+static constexpr int64_t PSTRG_RESERVED_RANGE_SIZE = 10000000;
 
 #define IS_IN_KEYRANGE(key, range) \
 	(key >= PSTRG_##range##_START && ((key - PSTRG_##range##_START) <= PSTRG_##range##_SIZE))
